EncryptoZ
I have completed my final year project EncryptoZ  which is a Digital Encipher System.It  is a framework that makes our data secure and guarantee the data storage security. When client uploads some data to the server then client must encrypt the data so that unauthenticated person can’t access it After uploading some data to the server, the server encrypt the data for storing/ transmitting the data. When client demands data, then server decrypts the data and client download it Therefore, under the aegis of this project it is intended to develop framework to avoid Data Theft, avoid Data Loss, facilitate easy on-demand self-service and cater to strength requirements